requirements.txt
langchain
openai
huggingface_hub
python-dotenv
streamlit
Ipython
google-generativeai
